feat: restructure task components and add modals for date and priority management
- Refactored TaskCard to import TaskCard from the new task directory structure.
- Added DatePickerModal for changing task due dates.
- Introduced PriorityModal for setting task priority levels.
- Created SubtaskProgress component to display progress of subtasks.
- Implemented TaskMeta for displaying task metadata including category, due date, and priority.
- Developed TaskMenu for task actions (Edit, Move to, Delete).
- Added README.md for task components documentation.
- Updated index.js for barrel exports of task components.
