+++ frontend/src/components/task/PriorityModal.jsx
@@ -1,0 +1,108 @@
+import React from 'react';
+import { motion, AnimatePresence } from 'framer-motion';
+import { Flag, X } from 'lucide-react';
+
+const PriorityModal = ({ 
+  isOpen, 
+  onClose, 
+  task, 
+  onPriorityChange, 
+  onEdit, 
+  isDark 
+}) => {
+  return (
+    <AnimatePresence>
+      {isOpen && (
+        <motion.div 
+          className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center p-4 z-50"
+          initial={{ opacity: 0 }}
+          animate={{ opacity: 1 }}
+          exit={{ opacity: 0 }}
+        >
+          <motion.div 
+            className={`${
+              isDark ? 'bg-gray-800 border-gray-700' : 'bg-white border-gray-200'
+            } rounded-2xl shadow-2xl border max-w-sm w-full p-6`}
+            initial={{ scale: 0.9, opacity: 0 }}
+            animate={{ scale: 1, opacity: 1 }}
+            exit={{ scale: 0.9, opacity: 0 }}
+            transition={{ duration: 0.2 }}
+          >
+            <div className="flex items-center justify-between mb-4">
+              <h3 className={`text-lg font-semibold ${isDark ? 'text-white' : 'text-gray-900'}`}>
+                Change Priority
+              </h3>
+              <button
+                onClick={onClose}
+                className={`p-2 rounded-lg transition-colors duration-200 ${
+                  isDark
+                    ? 'text-gray-400 hover:text-white hover:bg-gray-700'
+                    : 'text-gray-500 hover:text-gray-700 hover:bg-gray-100'
+                }`}
+              >
+                <X className="w-5 h-5" />
+              </button>
+            </div>
+            
+            <div className="space-y-4">
+              <div>
+                <label className={`block text-sm font-medium mb-3 ${isDark ? 'text-gray-300' : 'text-gray-700'}`}>
+                  <Flag className="w-4 h-4 inline mr-1" />
+                  Priority Level
+                </label>
+                <div className="grid grid-cols-3 gap-2">
+                  {[
+                    { value: 'low', label: 'Low', color: isDark ? 'bg-green-900/30 text-green-300 border-green-800' : 'bg-green-100 text-green-800 border-green-200' },
+                    { value: 'medium', label: 'Medium', color: isDark ? 'bg-yellow-900/30 text-yellow-300 border-yellow-800' : 'bg-yellow-100 text-yellow-800 border-yellow-200' },
+                    { value: 'high', label: 'High', color: isDark ? 'bg-red-900/30 text-red-300 border-red-800' : 'bg-red-100 text-red-800 border-red-200' }
+                  ].map(({ value, label, color }) => (
+                    <button
+                      key={value}
+                      onClick={() => {
+                        onPriorityChange?.(task, value) || onEdit({ ...task, priority: value });
+                        onClose();
+                      }}
+                      className={`px-4 py-3 rounded-lg border text-sm font-medium transition-all duration-200 ${
+                        task.priority === value ? color : isDark ? 'bg-gray-700 text-gray-300 border-gray-600 hover:bg-gray-600' : 'bg-gray-50 text-gray-600 border-gray-200 hover:bg-gray-100'
+                      }`}
+                    >
+                      {label}
+                    </button>
+                  ))}
+                </div>
+              </div>
+              
+              <div className="flex gap-3">
+                <button
+                  onClick={onClose}
+                  className={`flex-1 px-4 py-2 rounded-lg border transition-all duration-200 ${
+                    isDark
+                      ? 'border-gray-600 text-gray-300 hover:bg-gray-700'
+                      : 'border-gray-300 text-gray-600 hover:bg-gray-50'
+                  }`}
+                >
+                  Cancel
+                </button>
+                <button
+                  onClick={() => {
+                    onPriorityChange?.(task, null) || onEdit({ ...task, priority: null });
+                    onClose();
+                  }}
+                  className={`flex-1 px-4 py-2 rounded-lg transition-all duration-200 ${
+                    isDark
+                      ? 'bg-gray-700 text-gray-300 hover:bg-gray-600 border-gray-600'
+                      : 'bg-gray-100 text-gray-600 hover:bg-gray-200 border-gray-200'
+                  } border`}
+                >
+                  Remove Priority
+                </button>
+              </div>
+            </div>
+          </motion.div>
+        </motion.div>
+      )}
+    </AnimatePresence>
+  );
+};
+
+export default PriorityModal;
