+++ frontend/src/components/task/index.js
@@ -1,0 +1,6 @@
+export { default as TaskCard } from './TaskCard';
+export { default as TaskMenu } from './TaskMenu';
+export { default as DatePickerModal } from './DatePickerModal';
+export { default as PriorityModal } from './PriorityModal';
+export { default as TaskMeta } from './TaskMeta';
+export { default as SubtaskProgress } from './SubtaskProgress';
